% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_fisheye.R
\name{camera_fisheye}
\alias{camera_fisheye}
\title{Provide circular mask parameters from known camera+fisheye lens models}
\usage{
camera_fisheye(model = NULL)
}
\arguments{
\item{model}{Character. An input camera+lens model}
}
\value{
A list of three parameters (xc, yc, rc) of the circular mask
}
\description{
Provide circular mask parameters from known camera+fisheye lens models
}
\examples{
#available camera+lenses:
list.cameras

camera_fisheye(model='Coolpix4500+FC-E8')
}
