% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemispheR-package.R
\docType{package}
\name{hemispheR-package}
\alias{hemispheR}
\alias{hemispheR-package}
\title{hemispheR: Processing Hemispherical Canopy Images}
\description{
Import and classify canopy fish-eye images, estimate angular gap fraction and derive canopy attributes like leaf area index and openness. Additional information is provided in the study by Chianucci F., Macek M. (2023) \doi{10.1016/j.agrformet.2023.109470}.
}
\author{
\strong{Maintainer}: Francesco Chianucci \email{fchianucci@gmail.com} (\href{https://orcid.org/0000-0002-5688-2060}{ORCID})

Authors:
\itemize{
  \item Martin Macek \email{martin.macek@volny.cz} (\href{https://orcid.org/0000-0002-5609-5921}{ORCID})
}

}
\keyword{internal}
